Correção da atividade

+++ Tarefa_Wrapper/src/variaveis.java
@@ -2,7 +2,7 @@ public class variaveis {
     public static void main(String[] args) {
         int idade = 18;
         System.out.println(idade);
-        long idadeNova = idade;
+        Integer idadeNova = idade;
         System.out.println(idadeNova);
     }
 }
